ExploreResponse: Objects are a list of objects, not a single one.
Before if it was a list of objects, the first was used as parent object,
which was wrong. As a result if a list of objects was in the response
(without a parent object), the first object was not processed correct
(as a result, exploring the list of DBs, the first DB was missing.
The change thus affects all other parts which use the explore function.