Use DeleteObject for the session-id before disconnect the socket.
Before disconnecting, the object for the session-id should be deleted.
Also save the 2nd session-id we get at the connection setup, for further
usage (e.g. alarms, subscriptions).
The DeleteObject response has no integrity-id in the response, if we
delete our own session. It gives a warning (I suspect) in the return
value, but no fatal error.