Add threadsafe queue for received S7-PDUs.
When using functions with notification like subscriptions or alarming,
we may receive a new PDU when the current is still not processed
completely. As the receive is done via S7client from another thread, we
need to add and delete the PDUs to a queue in a thread safe manner. The
global error value as used now wouldn't work as expected in all cases.
In general this needs some bigger redesign when working dealing with
asynchronous protocol data like notifications.