Add Subscriptions and Alarm handling
Both methods are functional, but it's more to find out of how to use
this type of functions, and how to integrate them into the library.
Subcriptions and alarms use the same ressource, so for now they can't be
used parallel. Maybe it's possible to create a subscription for both,
variable subscriptions and alarms. But maybe it's easier to manage when
creating a separate connection e.g. for alarms.